refactor: 디렉토리 이동 및 클래스명 수정
