+++ src/main/kotlin/com/ssak3/timeattack/member/repository/MemberRepositoryCustom.kt
@@ -1,11 +1,11 @@
 package com.ssak3.timeattack.member.repository
 
-import com.ssak3.timeattack.member.domain.Member
-import com.ssak3.timeattack.member.domain.OAuthProvider
+import com.ssak3.timeattack.member.repository.entity.MemberEntity
+import com.ssak3.timeattack.member.repository.entity.OAuthProvider
 
 interface MemberRepositoryCustom {
     fun findByProviderAndSubject(
         oauthProvider: OAuthProvider,
         subject: String,
-    ): Member?
+    ): MemberEntity?
 }
