+++ src/main/kotlin/com/ssak3/timeattack/member/repository/MemberRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.ssak3.timeattack.member.repository
 
 import com.querydsl.jpa.impl.JPAQueryFactory
-import com.ssak3.timeattack.member.domain.Member
-import com.ssak3.timeattack.member.domain.OAuthProvider
 import com.ssak3.timeattack.member.domain.QMember
+import com.ssak3.timeattack.member.repository.entity.MemberEntity
+import com.ssak3.timeattack.member.repository.entity.OAuthProvider
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -13,7 +13,7 @@ class MemberRepositoryImpl(
     override fun findByProviderAndSubject(
         oauthProvider: OAuthProvider,
         subject: String,
-    ): Member? {
+    ): MemberEntity? {
         val qMember = QMember.member
         return queryFactory
             .select(qMember)
