+++ src/main/kotlin/com/ssak3/timeattack/task/domain/Task.kt
@@ -1,6 +1,6 @@
 package com.ssak3.timeattack.task.domain
 
-import com.ssak3.timeattack.member.domain.Member
+import com.ssak3.timeattack.member.repository.entity.MemberEntity
 import com.ssak3.timeattack.persona.domain.Persona
 import java.time.LocalDateTime
 
@@ -12,7 +12,7 @@ class Task(
     val triggerAction: String? = null,
     val estimatedTime: Int? = null,
     val status: TaskStatus,
-    val member: Member,
+    val member: MemberEntity,
     val persona: Persona,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
