+++ src/main/kotlin/com/ssak3/timeattack/task/repository/entity/TaskEntity.kt
@@ -1,7 +1,7 @@
 package com.ssak3.timeattack.task.repository.entity
 
 import com.ssak3.timeattack.common.domain.BaseEntity
-import com.ssak3.timeattack.member.domain.Member
+import com.ssak3.timeattack.member.repository.entity.MemberEntity
 import com.ssak3.timeattack.persona.repository.entity.PersonaEntity
 import com.ssak3.timeattack.task.domain.TaskCategory
 import com.ssak3.timeattack.task.domain.TaskStatus
@@ -40,7 +40,7 @@ class TaskEntity(
     val status: TaskStatus,
     @ManyToOne
     @JoinColumn(name = "member_id")
-    val member: Member,
+    val member: MemberEntity,
     // TODO: MemberEntity로 변경
     @ManyToOne
     @JoinColumn(name = "persona_id")
